v6.7.8

+++ released/MyBox/src/main/java/mara/mybox/controller/ControlData2DList.java
@@ -24,7 +24,8 @@ import mara.mybox.db.DerbyBase;
 import mara.mybox.db.data.Data2DDefinition;
 import mara.mybox.db.table.TableData2DDefinition;
 import mara.mybox.dev.MyBoxLog;
-import mara.mybox.fxml.SingletonCurrentTask;
+import mara.mybox.fxml.FxSingletonTask;
+import mara.mybox.fxml.FxTask;
 import mara.mybox.fxml.cell.TableDateCell;
 import mara.mybox.fxml.cell.TableNumberCell;
 import mara.mybox.fxml.style.StyleTools;
@@ -133,7 +134,7 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
 
     public void loadList() {
         if (manageController instanceof MyBoxTablesController) {
-            task = new SingletonCurrentTask<Void>(this) {
+            task = new FxSingletonTask<Void>(this) {
 
                 @Override
                 protected boolean handle() {
@@ -159,7 +160,7 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
             start(task);
 
         } else if (manageController instanceof DataTablesController) {
-            task = new SingletonCurrentTask<Void>(this) {
+            task = new FxSingletonTask<Void>(this) {
 
                 @Override
                 protected boolean handle() {
@@ -276,7 +277,7 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
     }
 
     @Override
-    protected int deleteData(List<Data2DDefinition> data) {
+    protected int deleteData(FxTask currentTask, List<Data2DDefinition> data) {
         if (data == null || data.isEmpty()) {
             return 0;
         }
@@ -286,6 +287,9 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
             try (Connection conn = DerbyBase.getConnection();
                     Statement statement = conn.createStatement()) {
                 for (Data2DDefinition item : data) {
+                    if (currentTask == null || !currentTask.isWorking()) {
+                        break;
+                    }
                     if (item.isUserTable() && item.getSheet() != null) {
                         String referName = DerbyBase.fixedIdentifier(item.getSheet());
                         try {
@@ -325,7 +329,7 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
     }
 
     @Override
-    protected long clearData() {
+    protected long clearData(FxTask currentTask) {
         if (manageController instanceof Data2DManageController
                 || manageController instanceof DataTablesController) {
             String sql = "SELECT d2did, sheet FROM " + tableData2DDefinition.getTableName()
@@ -337,18 +341,27 @@ public class ControlData2DList extends BaseSysTableController<Data2DDefinition> 
                     ResultSet results = query.executeQuery(sql)) {
                 List<String> names = new ArrayList<>();
                 while (results.next()) {
+                    if (currentTask == null || !currentTask.isWorking()) {
+                        break;
+                    }
                     names.add(results.getString("sheet"));
                     isCurrent = manageController.data2D != null
                             && results.getLong("d2did") == manageController.data2D.getD2did();
                 }
-                for (String name : names) {
-                    String tname = DerbyBase.fixedIdentifier(name);
-                    try {
-                        delete.executeUpdate("DROP TABLE " + tname);
-                    } catch (Exception e) {
-                        MyBoxLog.debug(e, tname);
+                if (currentTask != null && currentTask.isWorking()) {
+                    for (String name : names) {
+                        if (currentTask == null || !currentTask.isWorking()) {
+                            break;
+                        }
+                        String tname = DerbyBase.fixedIdentifier(name);
+                        try {
+                            delete.executeUpdate("DROP TABLE " + tname);
+                        } catch (Exception e) {
+                            MyBoxLog.debug(e, tname);
+                        }
                     }
                 }
+
             } catch (Exception e) {
                 MyBoxLog.error(e);
                 return -1;
